Added Like Function

+++ moviefansapp/main/models.py
@@ -35,7 +35,7 @@ class Movie(models.Model):
     genre_id = models.ForeignKey(Genre, on_delete=models.CASCADE)
     description = models.CharField(max_length=CONTENT_MAX_LENGTH)
     thumbnail = models.ImageField(upload_to="movie_thumbnails/", blank=True)
-    rating = models.FloatField()
+    rating = models.IntegerField()  # likes
     year = models.IntegerField()
     slug = models.SlugField(unique=True)
     views = models.IntegerField()
